+++ GameObjects/Obstacles/Bumper.h
@@ -12,20 +12,20 @@ using namespace std;
 class Bumper : public Obstacle{
 private:
     Vector2D position;
-    float radius;
+    double radius;
     int bscore;
     BumperType type;
 public:
-    Bumper(Vector2D pos, float rad, BumperType type);
+    Bumper(Vector2D pos, double rad, BumperType type);
 
     //Getters
     Vector2D GetPosition();
-    float GetRadius() ;
+    double GetRadius() ;
     BumperType GetType();
 
     //Setters
     void setScorepoints(int x);
 
-    Vector2D collideWith(Ball& ball, float collision_time, Manager & manager) override;
+    Vector2D collideWith(Ball& ball, double collision_time, Manager & manager) override;
     void draw(Interface & interface) override;
 };
