+++ GameObjects/Obstacles/Lane.cpp
@@ -22,7 +22,7 @@ void Lane::draw(Interface &interface) {
     interface.drawLane( Length, Center, Width);
 }
 
-Vector2D Lane::collideWith(Ball &ball, float collision_time, Manager &manager) {
+Vector2D Lane::collideWith(Ball &ball, double collision_time, Manager &manager) {
     bool Bottom= BICCollision(ball,Vector2D {Center.x,Center.y+Length/2.0},Width),
     Top=  BICCollision(ball,Vector2D {Center.x,Center.y-Length/2.0},Width);
