+++ GameObjects/Obstacles/Lane.h
@@ -20,6 +20,6 @@ public:
     Vector2D GetCenter() const;
 
     void draw(Interface& interface)override ;
-    Vector2D collideWith(Ball & ball, float collision_time, Manager & manager) override;
+    Vector2D collideWith(Ball & ball, double collision_time, Manager & manager) override;
 };
