+++ GameObjects/Obstacles/Magnet.h
@@ -20,6 +20,6 @@ public:
     void setMagnetRadius(double magnetRadius);
     void setRadius(double radius);
 
-    Vector2D collideWith(Ball & ball, float collision_time, Manager & manager)override;
+    Vector2D collideWith(Ball & ball, double collision_time, Manager & manager)override;
     void draw(Interface & interface) override;
 };
