+++ GameObjects/Obstacles/SpeedBooster.h
@@ -5,12 +5,12 @@
 
 class SpeedBooster :public Obstacle{
 private:
-    float radius;
+    double radius;
     Vector2D Center;
 public:
-    SpeedBooster(float radius, Vector2D Center);
+    SpeedBooster(double radius, Vector2D Center);
     void draw(Interface & interface) override;
-    Vector2D collideWith(Ball& ball, float collision_time, Manager & manager) override;
+    Vector2D collideWith(Ball& ball, double collision_time, Manager & manager) override;
 };
