+++ GameObjects/Obstacles/Switch.cpp
@@ -10,6 +10,6 @@ void Switch::draw(Interface& interface)
 	interface.drawSwitch(position);
 }
 
-Vector2D Switch::collideWith(Ball &ball, float collision_time, Manager &manager) {
+Vector2D Switch::collideWith(Ball &ball, double collision_time, Manager &manager) {
     return {0,0};
 }
