+++ GameObjects/Obstacles/ThrustBumper.cpp
@@ -1,6 +1,6 @@
 #include"ThrustBumper.h"
 
-ThrustBumper::ThrustBumper(Vector2D pos, float rad) : Bumper(pos, rad, BumperType::THRUST) {
+ThrustBumper::ThrustBumper(Vector2D pos, double rad) : Bumper(pos, rad, BumperType::THRUST) {
     setScorepoints(50);
 }
 
@@ -8,6 +8,6 @@ void ThrustBumper::draw(Interface& interface) {
     Bumper::draw(interface);
 }
 
-Vector2D ThrustBumper::collideWith(Ball& ball, float collision_time, Manager& manager) {
+Vector2D ThrustBumper::collideWith(Ball& ball, double collision_time, Manager& manager) {
     return Bumper::collideWith(ball, collision_time, manager);
 }
