+++ GameObjects/Obstacles/ThrustBumper.h
@@ -4,11 +4,10 @@
 
 class ThrustBumper : public Bumper {
 private:
-
 public:
-    ThrustBumper(Vector2D pos, float rad);
+    ThrustBumper(Vector2D pos, double rad);
     void draw(Interface& interface) override;
-    Vector2D collideWith(Ball& ball, float collision_time, Manager& manager)override;
+    Vector2D collideWith(Ball& ball, double collision_time, Manager& manager)override;
 };
