+++ GameObjects/Obstacles/VibraniumBumper.h
@@ -6,9 +6,9 @@ class VibraniumBumper : public Bumper {
 private:
 
 public:
-    VibraniumBumper(Vector2D pos, float rad);
+    VibraniumBumper(Vector2D pos, double rad);
     void draw(Interface& interface) override;
-    Vector2D collideWith(Ball& ball, float collision_time, Manager& manager)override;
+    Vector2D collideWith(Ball& ball, double collision_time, Manager& manager)override;
 };
